SHOP-5 | 에러 케이스 처리 및 Readme 추가

+++ src/main/java/com/shop/display/application/BrandService.java
@@ -2,6 +2,7 @@ package com.shop.display.application;
 
 import com.shop.display.domain.Brand;
 import com.shop.display.domain.BrandRepository;
+import com.shop.display.domain.ProductRepository;
 import com.shop.display.interfaces.exception.BadRequestException;
 import com.shop.display.interfaces.model.BrandCommand;
 import jakarta.transaction.Transactional;
@@ -10,17 +11,20 @@ import org.springframework.stereotype.Service;
 @Service
 public class BrandService {
     private final BrandRepository brandRepository;
+    private final ProductRepository productRepository;
 
-    public BrandService(BrandRepository brandRepository) {
+    public BrandService(BrandRepository brandRepository,
+                        ProductRepository productRepository) {
         this.brandRepository = brandRepository;
+        this.productRepository = productRepository;
     }
 
     @Transactional
     public void save(BrandCommand command) {
         brandRepository.findBrandByName(command.name())
-                          .ifPresent(brand -> {
-                              throw new BadRequestException("Brand already exists");
-                          });
+                       .ifPresent(brand -> {
+                           throw new BadRequestException("Brand already exists");
+                       });
         brandRepository.save(command.name());
     }
 
@@ -34,8 +38,12 @@ public class BrandService {
 
     @Transactional
     public void delete(long id) {
-        brandRepository.findBrandById(id)
-                       .orElseThrow(() -> new BadRequestException("Brand not found"));
+        Brand brand = brandRepository.findBrandById(id)
+                                     .orElseThrow(() -> new BadRequestException("Brand not found"));
+        boolean productsExistByBrandId = productRepository.isProductsExistByBrandId(brand);
+        if (productsExistByBrandId) {
+            throw new BadRequestException("Products exist by brand");
+        }
         brandRepository.delete(id);
     }
 }
